Rename all the Docstore classes
